modules/speech_to_text/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Async Tasks
celery==5.3.6
redis==5.0.1

# Speech-to-Text Providers
openai-whisper==20231117
google-cloud-speech==2.24.0
boto3==1.34.34

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
pyaudio==0.2.14

# Speaker Diarization
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2

# UI
streamlit==1.30.0
pandas==2.2.0

# Utilities
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Optional: For GPU acceleration
# torch==2.1.2+cu118
# torchaudio==2.1.2+cu118
